requirements.txt
appnope==0.1.0
asgiref==3.2.9
astroid==2.4.2
autopep8==1.5.3
Babel==2.8.0
Brotli==1.0.7
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.40.0
django-filter==2.3.0
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-twilio==0.13.0.post1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
googletrans==3.0.0
gunicorn==20.0.4
idna==2.9
infobip
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jsonpickle==1.4.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
phonenumbers==8.12.5
pickleshare==0.7.5
pika==1.1.0
Pillow==7.1.2
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
twilio==6.42.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
